Preparing for GitHub migration
As the project move towards GitHub, many of the tools used in GitLab are
deprecated. Here we have a few changes to the Cargo.toml and README.md
files along with base CI/CD for GitHub (without correct code coverage
showing up as GitHub does not provide an included tool).